WIP: Split plotting backends, default to Plotly
Addresses #338.

Now trying to import OrbitPlotter directly from poliastro.plotting
fails, because it's the "old" backend and users are encouraged to
use OrbitPlotter2D and OrbitPlotter3D instead, which use plotly
and not matplotlib behind the scenes.

This does not address:

* Updates in documentation
* Failing test because OrbitPlotter2D has no set_frame method (#480)
* Test simplification and refactoring (#448 and #447)

Also, notice that #281 should be fixed first.
